Mrgreentea/feat/35 update calendar events (#52)
* chore(app): 35 remove silvester, add mitmach call

* test(tools): different port for astro dev

This is to mimick the behaviour by SvelteKit. This allows leaving `dev` running while running the tests for example.

* feat(app): 35 Add info link for events to calendar

Some events have a custom page for them. These are now linked to when the event is selected in the calendar view.

Closes #36

* perf(app): 35 fix calendar stutter

When clicking on an event sometimes the browser would hang for a second. This fixes it by using svelte transitions.