feat(app): #38 update posts from instagram API (#78)
* feat(app): #38 update posts from instagram API

This might break at any time if instagram changes the query_id. But the alternative is setting up an OAuth app which I couldn't get to work and might not be worth the hassle (at the moment).

Closes: #38

* test(website): #38 use hard coded insta posts in tests

* perf(website): #33 Correct Image optimizations for Instagram posts

Because of the h-52 w-52 clases on the Image component we know the image will always be 13rem wide. We now use this information in the sizes attribute and optmize the images accordingly.

* refactor(website): #38 Make count of insta posts easy to change

Just added a count argument to the getRecentPhotos function.